requirements.txt
beautifulsoup4>=4.12,<5
python-docx>=1.1,<2
requests>=2.32,<3
